Refactored main plugin class & unregistered listeners on server stop
